test(xtream): split the live-stream layout spec under the max-lines cap
